Applying MVP Pattern
